Move: move pure go code to robotgo_pub.go
